Deprecate quantity description classes
