+++ editor/frontend/src/ProjectLocation/ProjectGeolocation.ts
@@ -6,11 +6,11 @@
  * @module Editing
  */
 
-import { AccuDrawHintBuilder, BeButtonEvent, CanvasDecoration, CoreTools, createQuantityDescription, DecorateContext, EventHandled, GraphicType, IModelApp, PrimitiveTool, ToolAssistance, ToolAssistanceImage, ToolAssistanceInputMethod, ToolAssistanceInstruction, ToolAssistanceSection, Viewport } from "@itwin/core-frontend";
+import { AccuDrawHintBuilder, BeButtonEvent, CanvasDecoration, CoreTools, DecorateContext, EventHandled, GraphicType, IModelApp, PrimitiveTool, ToolAssistance, ToolAssistanceImage, ToolAssistanceInputMethod, ToolAssistanceInstruction, ToolAssistanceSection, Viewport } from "@itwin/core-frontend";
 import { Angle, Matrix3d, Point3d, Ray3d, Vector3d, XYAndZ } from "@itwin/core-geometry";
 import { Cartographic, ColorDef, LinePixels } from "@itwin/core-common";
 import { ProjectExtentsClipDecoration } from "./ProjectExtentsDecoration";
-import { DialogItem, DialogProperty, DialogPropertySyncItem } from "@itwin/appui-abstract";
+import { type CustomFormattedNumberParams, DialogItem, DialogProperty, DialogPropertySyncItem, type PropertyDescription, PropertyEditorParamTypes, StandardEditorNames, StandardTypeNames } from "@itwin/appui-abstract";
 import { EditTools } from "../EditTool";
 
 function translatePrompt(key: string) {
@@ -18,6 +18,38 @@ function translatePrompt(key: string) {
 }
 function translateMessage(key: string) {
   return EditTools.translate(`ProjectLocation:Message.${key}`);
+}
+
+interface QuantityPropertyDescriptionProps {
+  name: string;
+  displayLabel: string;
+  kindOfQuantityName: string;
+  persistenceUnitName: string;
+  parseError: string;
+}
+
+function createQuantityPropertyDescription(props: QuantityPropertyDescriptionProps): PropertyDescription {
+  const { name, displayLabel, kindOfQuantityName, persistenceUnitName, parseError } = props;
+  const formatSpecHandle = IModelApp.quantityFormatter.getFormatSpecHandle(kindOfQuantityName, persistenceUnitName);
+  const editorParams: CustomFormattedNumberParams[] = [{
+    type: PropertyEditorParamTypes.CustomFormattedNumber,
+    formatFunction: (numberValue: number): string => formatSpecHandle.format(numberValue),
+    parseFunction: (userInput: string) => {
+      const parseResult = formatSpecHandle.parserSpec?.parseToQuantityValue(userInput);
+      return parseResult?.ok ? { value: parseResult.value } : { parseError };
+    },
+  }];
+
+  return {
+    name,
+    displayLabel,
+    kindOfQuantityName,
+    typename: StandardTypeNames.Number,
+    editor: {
+      name: StandardEditorNames.NumberCustom,
+      params: editorParams,
+    },
+  };
 }
 
 /** @internal */
@@ -102,7 +134,7 @@ export class ProjectGeolocationPointTool extends PrimitiveTool {
   private _latitudeProperty: DialogProperty<number> | undefined;
   public get latitudeProperty() {
     if (!this._latitudeProperty)
-      this._latitudeProperty = new DialogProperty<number>(createQuantityDescription({
+      this._latitudeProperty = new DialogProperty<number>(createQuantityPropertyDescription({
         name: "latitude", displayLabel: translateMessage("Latitude"),
         kindOfQuantityName: "DefaultToolsUnits.ANGLE", persistenceUnitName: "Units.RAD",
         parseError: IModelApp.localization.getLocalizedString("iModelJs:Properties.UnableToParseAngle"),
@@ -116,7 +148,7 @@ export class ProjectGeolocationPointTool extends PrimitiveTool {
   private _longitudeProperty: DialogProperty<number> | undefined;
   public get longitudeProperty() {
     if (!this._longitudeProperty)
-      this._longitudeProperty = new DialogProperty<number>(createQuantityDescription({
+      this._longitudeProperty = new DialogProperty<number>(createQuantityPropertyDescription({
         name: "longitude", displayLabel: translateMessage("Longitude"),
         kindOfQuantityName: "DefaultToolsUnits.ANGLE", persistenceUnitName: "Units.RAD",
         parseError: IModelApp.localization.getLocalizedString("iModelJs:Properties.UnableToParseAngle"),
@@ -130,7 +162,7 @@ export class ProjectGeolocationPointTool extends PrimitiveTool {
   private _altitudeProperty: DialogProperty<number> | undefined;
   public get altitudeProperty() {
     if (!this._altitudeProperty)
-      this._altitudeProperty = new DialogProperty<number>(createQuantityDescription({
+      this._altitudeProperty = new DialogProperty<number>(createQuantityPropertyDescription({
         name: "altitude", displayLabel: CoreTools.translate("Measure.Labels.Altitude"),
         kindOfQuantityName: "DefaultToolsUnits.LENGTH", persistenceUnitName: "Units.M",
         parseError: IModelApp.localization.getLocalizedString("iModelJs:Properties.UnableToParseLength"),
@@ -144,7 +176,7 @@ export class ProjectGeolocationPointTool extends PrimitiveTool {
   private _northProperty: DialogProperty<number> | undefined;
   public get northProperty() {
     if (!this._northProperty)
-      this._northProperty = new DialogProperty<number>(createQuantityDescription({
+      this._northProperty = new DialogProperty<number>(createQuantityPropertyDescription({
         name: "north", displayLabel: translateMessage("North"),
         kindOfQuantityName: "DefaultToolsUnits.ANGLE", persistenceUnitName: "Units.RAD",
         parseError: IModelApp.localization.getLocalizedString("iModelJs:Properties.UnableToParseAngle"),
