+++ example-code/snippets/package.json
@@ -21,7 +21,6 @@
   "repository": {},
   "dependencies": {
     "@azure/storage-blob": "^12.28.0",
-    "@itwin/appui-abstract": "workspace:*",
     "@itwin/core-backend": "workspace:*",
     "@itwin/core-bentley": "workspace:*",
     "@itwin/core-common": "workspace:*",
